Close all <a> tags
Self-closing tags with `/>` is invalid in the HTML5 world. For example, it's just `<br>` and not `<br />` anymore. But you need closing tags for anchors because a link has a beginning and an end (and anchors are used for both "anchoring" and "hyperlinking"). Made this PR because I got bothered by the weird way this package's documentation looks in the [atmosphere.js repository](https://atmospherejs.com/splendido/accounts-meld) :D